rename the new page
